Drop obsolete btattach udev rule + systemd service
Drop the obsolete udev rule + systemd service to call btattach on BT-HCIs
connected via UART from userspace.

I added these a long time ago to make BT work on Broadcom HCIs connected
through an UART, but this has not been necessary for a long time.

The kernel now makes the connection between the UART and the HCI itself
and as part of this it no longer instantiates the platform-devices which
the udev rule triggers on, so these have been a no-op for a long time now.