Add the btmgmt util to the packaged files
I have been debugging a regression where Bluetooth has stopped working on
some broadcom Wifi/BT chipsets. This is caused by the BT parts of these
chips not having a unique MAC programmed into them. Instead they all use
the same factory-default address. This used to work as long as only one
device with this factory-default address was in the room. But a while ago the
kernel learned to recognize the factory-default address and now it (rightly)
refuses to use this. Instead it expects userspace to configure a valid MAC
address.

Both checking if their are any adapters in the unconfigured state (these
are invisible in e.g. bluetoothctl) and setting their MAC address to make
them work requires the btmgmt util, so lets package it.

Note this is not a complete fix for this issue, but having btmgmt
available will at least allow users to work around the problem.

I've filed an upstream bluez issue for getting a workaround for this
in place inside bluez: https://github.com/bluez/bluez/issues/107